+++ Source/GxB_Vector_memoryUsage.c
@@ -29,7 +29,7 @@ GrB_Info GxB_Vector_memoryUsage  // return # of bytes used for a vector
     // get the memory size taken by the vector
     //--------------------------------------------------------------------------
 
-    int nallocs ;
+    int64_t nallocs ;
     size_t mem_shallow ;
     return (GB_memoryUsage (&nallocs, size, &mem_shallow, (GrB_Matrix) v)) ;
 }
